modularize client code
