+++ client/components/player.js
@@ -1,8 +1,8 @@
-const { LINK_HEIGHT, LINK_WIDTH, BORDER_MARGIN } = require('./enums/length');
-const { DIRECTION_RIGHT, DIRECTION_LEFT, DIRECTION_UP, DIRECTION_DOWN } = require('./enums/direction');
+const { LINK_HEIGHT, LINK_WIDTH, BORDER_MARGIN } = require('../enums/length');
+const { DIRECTION_RIGHT, DIRECTION_LEFT, DIRECTION_UP, DIRECTION_DOWN } = require('../enums/direction');
 
-const Link = require('./objects');
-const { ateFood, updateLinkLocation } = require('./utilities');
+const Link = require('../model/objects');
+const { ateFood, updateLinkLocation } = require('../utilities/utilities');
 
 function SnakePlayer(imageSrc, automated, ctx, initialPositionX, initialPositionY) {
     let startPositionX = !!initialPositionX ? initialPositionX : 0;
