--- client/enum.js
@@ -1,13 +1,0 @@
-const CANVAS_WIDTH = 920;
-const CANVAS_HEIGHT = 750;
-
-const LINK_WIDTH = 34;
-const LINK_HEIGHT = 34;
-
-const BORDER_MARGIN = 1;
-
-// Directions
-const DIRECTION_LEFT = "LEFT";
-const DIRECTION_RIGHT = "RIGHT";
-const DIRECTION_UP = "UP";
-const DIRECTION_DOWN = "DOWN";
