+++ client/utilities/utilities.js
@@ -1,5 +1,5 @@
-const { CANVAS_HEIGHT, CANVAS_WIDTH, LINK_HEIGHT, LINK_WIDTH, BORDER_MARGIN } = require('./enums/length');
-const { DIRECTION_RIGHT, DIRECTION_LEFT, DIRECTION_UP, DIRECTION_DOWN } = require('./enums/direction');
+const { CANVAS_HEIGHT, CANVAS_WIDTH, LINK_HEIGHT, LINK_WIDTH, BORDER_MARGIN } = require('../enums/length');
+const { DIRECTION_RIGHT, DIRECTION_LEFT, DIRECTION_UP, DIRECTION_DOWN } = require('../enums/direction');
 
 function getRndInteger(min, max) {
     return Math.floor(Math.random() * (max - min) ) + min;
